chore: move session migration helpers
